[DEV] Developing timeout events
EN: Developing timeout events. Now designing interfaces for lower level timeout service
CN: 开发超时事件，正在设计底层超时功能接口

+++ include/epEvent.h
@@ -52,7 +52,7 @@ struct AMCEpollEvent *
 
 /* Public Class Functions */
 struct AMCEpollEvent *
-	epEvent_New(int fd, events_t what, int timeout, ev_callback callback, void *userData);
+	epEvent_New(int fd, events_t what, long timeout, ev_callback callback, void *userData);
 int 
 	epEvent_Free(struct AMCEpollEvent *event);
 const char *
